feat: fishnet conversion
convert Sacados to fishnet which is more mature for now than netcode.

also add abstraction to the storage, with a local storage and a network one using fishnet. This would allow in the future the usage of other networking solutions (still require some changes to the serialization of the Item and ItemStack)